# Conflicts:
#	security/src/main/java/jm/JmAuthenticationSuccessHandler.java
#	security/src/main/java/jm/WebSecurityConfig.java
#	webapp/src/main/resources/application.properties
#	webapp/src/main/resources/templates/workspace-page.html

+++ webapp/src/main/resources/static/js/workspace-page/components/messages/StompClient.js
@@ -1,5 +1,5 @@
-// import {ChannelView} from "/js/workspace-page/components/sidebar/ChannelView.js";
 import {setOnClickEdit} from "/js/messagesInlineEdit.js";
+import {Command} from "/js/workspace-page/components/footer/Command.js";
 
 export class StompClient {
 
@@ -9,6 +9,8 @@ export class StompClient {
         this.thread_view = thread_view;
         this.dm_view = direct_message_view;
         this.channelview = channel_view;
+
+        this.commands = new Command();
 
         window.sendName = (message) => this.sendName(message);
         window.sendChannel = (channel) => this.sendChannel(channel);
@@ -29,8 +31,8 @@ export class StompClient {
     subscribeMessage() {
         this.stompClient.subscribe('/topic/messages', async (message) => {
             let result = JSON.parse(message.body);
+            result['content'] = result.inputMassage;
             if (result.userId != null && !result.isDeleted) {
-                result['content'] = result.inputMassage;
                 if (result.channelId === channel_id) {
                     if (result.isUpdated) {
                         this.channel_message_view.updateMessage(result);
@@ -44,15 +46,17 @@ export class StompClient {
                     this.channel_message_view.dialog.messageBoxWrapper();
                 }
             } else {
-                this.channel_message_view.dialog.deleteMessage(result.id, result.userId);
+                if (result.isDeleted) {
+                    this.channel_message_view.dialog.deleteMessage(result.id, result.userId);
+                } else {
+                    this.commands.checkMessage(result);
+                }
             }
             notifyParseMessage(result);
-            setOnClickEdit(true);
         });
     }
 
     subscribeChannel() {
-        // this.channelview = new ChannelView();
         this.stompClient.subscribe('/topic/channel', (channel) => {
             const chn = JSON.parse(channel.body);
             this.channelview.addChannelIntoSidebarChannelList(chn);
@@ -77,7 +81,6 @@ export class StompClient {
                     this.dm_view.updateMessage(response);
                 } else {
                     if (response.conversationId === current_conversation) {
-                        console.warn(response);
                         this.dm_view.createMessage(response);
                     }
                 }
